chore: replace styles with styled components
