--- styles/globals.css
@@ -1,29 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@200;400;600;700&display=swap");
-
-html,
-body {
-  background-color: #505050;
-  color: white;
-  padding: 0;
-  margin: 0;
-  font-family: "Montserrat", sans-serif;
-}
-
-h1 {
-  font-size: 3rem;
-  font-weight: 700;
-  margin: 0;
-}
-
-p {
-  font-weight: 200;
-}
-
-a {
-  color: inherit;
-  text-decoration: none;
-}
-
-* {
-  box-sizing: border-box;
-}
